Moves to latest zipkin reporter, dropping scribe
Scribe is a long-since deprecated transport and no longer maintained.
This updates to latest zipkin, and drops support for scribe.

Note: we should re-evaluate how to configure kafka (or rabbitmq) as they
have dependencies not everyone will use. For example, in spring-boot, we
have a compile dep on zipkin classes, but optional on the dependencies.